Do no cache or use cached hgnc downloads
It doesn't look like the cached download was ever being deleted
in the code, even after processing, so the hgnc updates were
only ever happening as a consequence of that file being manually
deleted.

Relatedly, when the cached file was there, the test related to this
functionality would use the whole cached file instead of the
smaller test version (causing the test to sometimes time out).